Add Hand Gesture Control Game Project (#1930)
## 🖐️ Hand Gesture Joystick Control

Fixes #1929

### 📌 What's Added
- Hand gesture controlled game using webcam
- Detects wrist movement to trigger arrow key presses (Up, Down, Left,
Right)
- Auto-calibrates neutral point on hand detection
- Press `c` to recalibrate, `q` to quit

### 🧠 Tech Stack
- OpenCV — Webcam capture
- MediaPipe — Hand landmark detection
- PyAutoGUI — Arrow key simulation

### 📁 Files Added
- `hand_gesture_controlled_game.ipynb` — Jupyter Notebook
- `README.md` — Project documentation

### 🎮 Demo Game
Can be tested with Subway Surfers → https://poki.com/en/g/subway-surfers
