+++ Gesture Related Projects/Real_Time_Hand_Gesture_Recognition_WebApp_Gestura_AI/requirements.txt
