+++ Gesture Related Projects/Recognition_of_gestures_and_action/image and video
