--- Real_Time_Hand_Gesture_Recognition_WebApp_Gestura_AI/fresh_gesture_app.py
@@ -1,823 +1,0 @@
-import os
-import cv2
-import time
-import joblib
-import numpy as np
-import streamlit as st
-import mediapipe as mp
-
-from streamlit_webrtc import webrtc_streamer, VideoProcessorBase
-from mediapipe.tasks import python
-from mediapipe.tasks.python import vision
-
-# =========================================================
-# PAGE CONFIG
-# =========================================================
-st.set_page_config(
-    page_title="Gestura AI · Hand Sign Recognition",
-    page_icon="🖐️",
-    layout="wide",
-    initial_sidebar_state="expanded"
-)
-
-# =========================================================
-# GLOBAL STYLES — Dark Editorial / Precision Tech Theme
-# =========================================================
-st.markdown("""
-<style>
-
-@import url('https://fonts.googleapis.com/css2?family=Space+Mono:wght@400;700&family=Outfit:wght@300;400;500;600;700;800&display=swap');
-
-/* ── Root Tokens ─────────────────────────────────── */
-:root {
-    --bg-base:      #070b14;
-    --bg-surface:   #0d1424;
-    --bg-raised:    #111c30;
-    --border:       rgba(255,255,255,0.07);
-    --border-glow:  rgba(0,230,200,0.35);
-    --accent-teal:  #00e6c8;
-    --accent-blue:  #3b82f6;
-    --accent-violet:#8b5cf6;
-    --text-primary: #f0f4ff;
-    --text-secondary:#94a3b8;
-    --text-dim:     #475569;
-    --green-live:   #22d3a0;
-    --red-warn:     #f87171;
-    --mono:         'Space Mono', monospace;
-    --sans:         'Outfit', sans-serif;
-    --radius-sm:    10px;
-    --radius-md:    16px;
-    --radius-lg:    24px;
-}
-
-/* ── Reset & Base ────────────────────────────────── */
-html, body, [class*="css"] {
-    font-family: var(--sans) !important;
-    color: var(--text-primary) !important;
-}
-
-.stApp {
-    background-color: var(--bg-base);
-    background-image:
-        radial-gradient(ellipse 80% 50% at 10% 5%,  rgba(59,130,246,0.08) 0%, transparent 60%),
-        radial-gradient(ellipse 60% 40% at 90% 90%, rgba(0,230,200,0.07) 0%, transparent 55%),
-        repeating-linear-gradient(
-            90deg,
-            transparent,
-            transparent 79px,
-            rgba(255,255,255,0.015) 80px
-        );
-}
-
-/* ── Header ─────────────────────────────────────── */
-header[data-testid="stHeader"] {
-    background: rgba(7,11,20,0.92) !important;
-    backdrop-filter: blur(16px);
-    border-bottom: 1px solid var(--border);
-}
-
-/* ── Sidebar ─────────────────────────────────────── */
-[data-testid="stSidebar"] {
-    background: var(--bg-surface) !important;
-    border-right: 1px solid var(--border);
-}
-[data-testid="stSidebar"] * {
-    color: var(--text-primary) !important;
-}
-[data-testid="stSidebar"] .stMarkdown p,
-[data-testid="stSidebar"] label,
-[data-testid="stSidebar"] span {
-    color: var(--text-secondary) !important;
-}
-
-/* ── Hero Title ─────────────────────────────────── */
-.hero-wrapper {
-    padding: 2.5rem 0 1.5rem;
-    border-bottom: 1px solid var(--border);
-    margin-bottom: 2rem;
-}
-.hero-eyebrow {
-    font-family: var(--mono);
-    font-size: 0.72rem;
-    letter-spacing: 0.22em;
-    text-transform: uppercase;
-    color: var(--accent-teal);
-    margin-bottom: 0.6rem;
-}
-.hero-title {
-    font-size: 3.2rem;
-    font-weight: 800;
-    line-height: 1.1;
-    color: var(--text-primary);
-    letter-spacing: -1.5px;
-    margin: 0 0 0.5rem;
-}
-.hero-title span {
-    background: linear-gradient(90deg, var(--accent-teal), var(--accent-blue));
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
-.hero-sub {
-    font-size: 1.0rem;
-    color: var(--text-secondary);
-    font-weight: 400;
-    max-width: 520px;
-}
-
-/* ── Panel Card ─────────────────────────────────── */
-.panel {
-    background: var(--bg-surface);
-    border: 1px solid var(--border);
-    border-radius: var(--radius-lg);
-    padding: 1.75rem;
-    position: relative;
-    overflow: hidden;
-}
-.panel::before {
-    content: "";
-    position: absolute;
-    top: 0; left: 0; right: 0;
-    height: 2px;
-    background: linear-gradient(90deg, var(--accent-teal), var(--accent-blue), var(--accent-violet));
-    border-radius: var(--radius-lg) var(--radius-lg) 0 0;
-}
-.panel-title {
-    font-family: var(--mono);
-    font-size: 0.68rem;
-    letter-spacing: 0.2em;
-    text-transform: uppercase;
-    color: var(--text-dim);
-    margin-bottom: 1.1rem;
-}
-
-/* ── Live Status Badge ──────────────────────────── */
-.live-badge {
-    display: inline-flex;
-    align-items: center;
-    gap: 7px;
-    background: rgba(34,211,160,0.1);
-    border: 1px solid rgba(34,211,160,0.25);
-    border-radius: 999px;
-    padding: 4px 12px;
-    font-family: var(--mono);
-    font-size: 0.72rem;
-    color: var(--green-live);
-    letter-spacing: 0.1em;
-    text-transform: uppercase;
-}
-.live-dot {
-    width: 8px; height: 8px;
-    border-radius: 50%;
-    background: var(--green-live);
-    animation: pulse-dot 1.4s ease-in-out infinite;
-}
-@keyframes pulse-dot {
-    0%, 100% { opacity: 1; transform: scale(1);   }
-    50%       { opacity: 0.5; transform: scale(1.5); }
-}
-
-/* ── Current Gesture Display ────────────────────── */
-.gesture-display {
-    background: linear-gradient(135deg, var(--bg-raised), var(--bg-surface));
-    border: 1px solid var(--border-glow);
-    border-radius: var(--radius-md);
-    padding: 1.5rem 2rem;
-    margin: 1rem 0;
-    text-align: center;
-}
-.gesture-emoji {
-    font-size: 3.5rem;
-    line-height: 1;
-    margin-bottom: 0.4rem;
-}
-.gesture-label {
-    font-family: var(--mono);
-    font-size: 1.1rem;
-    color: var(--accent-teal);
-    font-weight: 700;
-    letter-spacing: 0.06em;
-}
-.gesture-sub {
-    font-size: 0.8rem;
-    color: var(--text-dim);
-    margin-top: 0.25rem;
-    font-family: var(--mono);
-}
-
-/* ── Log Entry ───────────────────────────────────── */
-.log-entry {
-    display: flex;
-    align-items: center;
-    gap: 14px;
-    background: var(--bg-raised);
-    border: 1px solid var(--border);
-    border-radius: var(--radius-sm);
-    padding: 12px 16px;
-    margin-bottom: 8px;
-    transition: border-color 0.2s;
-}
-.log-entry:first-child {
-    border-color: rgba(0,230,200,0.3);
-    background: rgba(0,230,200,0.05);
-}
-.log-index {
-    font-family: var(--mono);
-    font-size: 0.65rem;
-    color: var(--text-dim);
-    min-width: 28px;
-}
-.log-gesture {
-    font-size: 1.05rem;
-    color: var(--text-primary);
-    font-weight: 500;
-    flex: 1;
-}
-.log-new-tag {
-    font-family: var(--mono);
-    font-size: 0.6rem;
-    letter-spacing: 0.15em;
-    text-transform: uppercase;
-    color: var(--accent-teal);
-    background: rgba(0,230,200,0.12);
-    border-radius: 4px;
-    padding: 2px 7px;
-}
-
-/* ── Metric Cards ────────────────────────────────── */
-.metric-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 12px;
-    margin-top: 0.5rem;
-}
-.metric-card {
-    background: var(--bg-raised);
-    border: 1px solid var(--border);
-    border-radius: var(--radius-md);
-    padding: 1.2rem;
-}
-.metric-label {
-    font-family: var(--mono);
-    font-size: 0.65rem;
-    text-transform: uppercase;
-    letter-spacing: 0.15em;
-    color: var(--text-dim);
-    margin-bottom: 0.5rem;
-}
-.metric-value {
-    font-family: var(--mono);
-    font-size: 2rem;
-    font-weight: 700;
-    color: var(--text-primary);
-    line-height: 1;
-}
-.metric-tag {
-    font-size: 0.78rem;
-    color: var(--accent-teal);
-    margin-top: 0.4rem;
-}
-
-/* ── Key Map Table ───────────────────────────────── */
-.keymap-row {
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    padding: 12px 16px;
-    border-radius: var(--radius-sm);
-    margin-bottom: 6px;
-    background: var(--bg-raised);
-    border: 1px solid var(--border);
-}
-.keymap-sign {
-    font-size: 1.6rem;
-    min-width: 40px;
-    text-align: center;
-}
-.keymap-name {
-    font-weight: 600;
-    font-size: 0.95rem;
-    color: var(--text-primary);
-    flex: 1;
-}
-.keymap-desc {
-    font-size: 0.78rem;
-    color: var(--text-secondary);
-    font-family: var(--mono);
-}
-
-/* ── Buttons ─────────────────────────────────────── */
-.stButton > button {
-    width: 100%;
-    background: transparent;
-    border: 1px solid var(--border-glow);
-    color: var(--accent-teal) !important;
-    border-radius: var(--radius-sm);
-    padding: 0.6rem 1.2rem;
-    font-family: var(--mono);
-    font-size: 0.75rem;
-    letter-spacing: 0.1em;
-    text-transform: uppercase;
-    transition: all 0.2s ease;
-    font-weight: 700;
-}
-.stButton > button:hover {
-    background: rgba(0,230,200,0.08);
-    border-color: var(--accent-teal);
-    box-shadow: 0 0 20px rgba(0,230,200,0.15);
-    transform: translateY(-1px);
-}
-
-/* ── Radio Tabs ─────────────────────────────────── */
-div[role="radiogroup"] {
-    gap: 8px;
-}
-div[role="radiogroup"] label {
-    background: var(--bg-raised) !important;
-    border: 1px solid var(--border) !important;
-    border-radius: var(--radius-sm) !important;
-    padding: 8px 16px !important;
-    font-family: var(--mono);
-    font-size: 0.72rem;
-    letter-spacing: 0.08em;
-    transition: all 0.2s;
-    color: var(--text-secondary) !important;
-}
-div[role="radiogroup"] label:hover {
-    border-color: var(--border-glow) !important;
-}
-div[role="radiogroup"] [aria-checked="true"] + label,
-div[role="radiogroup"] label[data-checked="true"] {
-    border-color: var(--accent-teal) !important;
-    color: var(--accent-teal) !important;
-}
-
-/* ── Info / Warning Boxes ────────────────────────── */
-.stAlert {
-    background: var(--bg-raised) !important;
-    border-radius: var(--radius-sm) !important;
-    border: 1px solid var(--border) !important;
-    color: var(--text-secondary) !important;
-}
-
-/* ── Dividers ────────────────────────────────────── */
-hr {
-    border: none;
-    border-top: 1px solid var(--border);
-    margin: 1.25rem 0;
-}
-
-/* ── Scrollbar ───────────────────────────────────── */
-::-webkit-scrollbar { width: 6px; }
-::-webkit-scrollbar-track { background: transparent; }
-::-webkit-scrollbar-thumb {
-    background: rgba(0,230,200,0.25);
-    border-radius: 10px;
-}
-
-/* ── Sidebar branding ────────────────────────────── */
-.sidebar-brand {
-    font-family: var(--mono);
-    font-size: 0.7rem;
-    text-transform: uppercase;
-    letter-spacing: 0.2em;
-    color: var(--text-dim) !important;
-    margin-bottom: 1.2rem;
-    padding-bottom: 1.2rem;
-    border-bottom: 1px solid var(--border);
-}
-.sidebar-stat {
-    display: flex;
-    justify-content: space-between;
-    padding: 8px 0;
-    border-bottom: 1px solid var(--border);
-    font-size: 0.82rem;
-}
-.sidebar-stat-label { color: var(--text-dim) !important; }
-.sidebar-stat-value {
-    font-family: var(--mono);
-    font-size: 0.78rem;
-    color: var(--accent-teal) !important;
-}
-
-</style>
-""", unsafe_allow_html=True)
-
-
-# =========================================================
-# MODEL ASSET LOADING
-# =========================================================
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-MODEL_DIR = os.path.join(BASE_DIR, "model", "keypoint_classifier")
-LANDMARKER_PATH = os.path.join(MODEL_DIR, "hand_landmarker.task")
-CLASSIFIER_PATH = os.path.join(MODEL_DIR, "gesture_classifier.pkl")
-
-@st.cache_resource
-def load_assets():
-    if not os.path.exists(CLASSIFIER_PATH) or not os.path.exists(LANDMARKER_PATH):
-        return None, None
-    clf = joblib.load(CLASSIFIER_PATH)
-    options = vision.HandLandmarkerOptions(
-        base_options=python.BaseOptions(model_asset_path=LANDMARKER_PATH),
-        running_mode=vision.RunningMode.IMAGE,
-        num_hands=1,
-        min_hand_detection_confidence=0.5
-    )
-    marker = vision.HandLandmarker.create_from_options(options)
-    return clf, marker
-
-model, landmarker = load_assets()
-
-
-# =========================================================
-# SESSION STATE — Thread-safe global containers
-# =========================================================
-if not hasattr(st, "global_gesture_queue"):
-    st.global_gesture_queue = []
-if not hasattr(st, "latest_detected_gesture"):
-    st.latest_detected_gesture = "No Hand Detected"
-# Store emoji and label separately — avoids fragile string splitting on the main thread
-if not hasattr(st, "latest_gesture_emoji"):
-    st.latest_gesture_emoji = "🫥"
-if not hasattr(st, "latest_gesture_label"):
-    st.latest_gesture_label = "No Hand Detected"
-if not hasattr(st, "total_detections"):
-    st.total_detections = 0
-if not hasattr(st, "session_start"):
-    st.session_start = time.time()
-
-
-# =========================================================
-# GESTURE LABELS
-# =========================================================
-GESTURE_NAMES = {
-    "0": "Open Palm",
-    "1": "Closed Fist",
-    "2": "Thumbs Up",
-    "3": "Peace Sign",
-    "4": "OK Sign",
-}
-
-GESTURE_EMOJI = {
-    "Open Palm":   "🖐️",
-    "Closed Fist": "✊",
-    "Thumbs Up":   "👍",
-    "Peace Sign":  "✌️",
-    "OK Sign":     "👌",
-}
-
-GESTURE_COLOR = {
-    "Open Palm":   (0, 230, 200),
-    "Closed Fist": (239, 68, 68),
-    "Thumbs Up":   (34, 211, 160),
-    "Peace Sign":  (99, 102, 241),
-    "OK Sign":     (251, 191, 36),
-}
-
-
-# =========================================================
-# VIDEO PROCESSOR
-# =========================================================
-class GestureProcessor(VideoProcessorBase):
-    def __init__(self):
-        self.last_prediction = None
-        self.stable_frame_count = 0
-        self.prev_time = time.time()
-        self.STABILITY_THRESHOLD = 4   # frames before committing a detection
-
-    def recv(self, frame):
-        img = frame.to_ndarray(format="bgr24")
-        img = cv2.flip(img, 1)
-        img = cv2.resize(img, (640, 480))
-        h, w, _ = img.shape
-
-        rgb_img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        mp_image = mp.Image(image_format=mp.ImageFormat.SRGB, data=rgb_img)
-        results = landmarker.detect(mp_image)
-
-        status_text = "Searching..."
-        dot_color   = (148, 163, 184)     # dim when idle
-        accent      = (0, 230, 200)
-
-        if results.hand_landmarks and model is not None:
-            hand = results.hand_landmarks[0]
-
-            # — Draw landmark skeleton —
-            for lm in hand:
-                cx, cy = int(lm.x * w), int(lm.y * h)
-                cv2.circle(img, (cx, cy), 4, accent, -1)
-                cv2.circle(img, (cx, cy), 7, (*accent, 80), 1)
-
-            # — Feature extraction (normalized relative to wrist) —
-            pts       = np.array([[lm.x, lm.y] for lm in hand])
-            shifted   = pts - pts[0]
-            scale     = max(np.abs(shifted).max(), 1e-6)
-            normalized = (shifted / scale).flatten().tolist()
-
-            raw_pred     = str(model.predict([normalized])[0])
-            gesture_name = GESTURE_NAMES.get(raw_pred, "Unknown")
-            dot_color    = GESTURE_COLOR.get(gesture_name, accent)
-            status_text  = gesture_name
-
-            emoji = GESTURE_EMOJI.get(gesture_name, "")
-            # Write emoji and label separately — main thread reads these directly
-            st.latest_gesture_emoji = emoji
-            st.latest_gesture_label = gesture_name
-            st.latest_detected_gesture = f"{emoji} {gesture_name}"  # kept for backward compat
-
-            # — Anti-jitter latch: only commit after stability threshold —
-            if raw_pred == self.last_prediction:
-                self.stable_frame_count += 1
-            else:
-                self.stable_frame_count = 0
-                self.last_prediction = raw_pred
-
-            if self.stable_frame_count >= self.STABILITY_THRESHOLD:
-                self.stable_frame_count = 0
-                display_name = f"{emoji} {gesture_name}"
-
-                # Deduplicate consecutive identical gestures
-                if (len(st.global_gesture_queue) == 0 or
-                        st.global_gesture_queue[-1] != display_name):
-                    print(f"[Gestura] Detected → {display_name}")
-                    st.global_gesture_queue.append(display_name)
-                    st.total_detections += 1
-        else:
-            st.latest_gesture_emoji = "🫥"
-            st.latest_gesture_label = "No Hand Detected"
-            st.latest_detected_gesture = "No Hand Detected"
-
-        # — FPS calculation —
-        now = time.time()
-        fps = int(1 / max(now - self.prev_time, 1e-6))
-        self.prev_time = now
-
-        # — HUD overlay —
-        overlay = img.copy()
-        cv2.rectangle(overlay, (0, 0), (w, 64), (7, 11, 20), -1)
-        cv2.addWeighted(overlay, 0.75, img, 0.25, 0, img)
-
-        # Status dot
-        cv2.circle(img, (24, 32), 7, dot_color, -1)
-
-        # Status text
-        cv2.putText(img, status_text,
-                    (42, 38), cv2.FONT_HERSHEY_DUPLEX,
-                    0.75, (240, 244, 255), 1, cv2.LINE_AA)
-
-        # FPS
-        fps_text = f"{fps} FPS"
-        fps_x    = w - cv2.getTextSize(fps_text, cv2.FONT_HERSHEY_SIMPLEX, 0.48, 1)[0][0] - 14
-        cv2.putText(img, fps_text,
-                    (fps_x, 38), cv2.FONT_HERSHEY_SIMPLEX,
-                    0.48, (71, 85, 105), 1, cv2.LINE_AA)
-
-        return frame.from_ndarray(img, format="bgr24")
-
-
-# =========================================================
-# SIDEBAR
-# =========================================================
-with st.sidebar:
-    st.markdown("""
-    <div class="sidebar-brand">Gestura AI &nbsp;·&nbsp; v2.0</div>
-    """, unsafe_allow_html=True)
-
-    st.markdown("**Session Overview**")
-
-    uptime_secs = int(time.time() - st.session_start)
-    uptime_str  = f"{uptime_secs // 60}m {uptime_secs % 60}s"
-
-    st.markdown(f"""
-    <div class="sidebar-stat">
-        <span class="sidebar-stat-label">Detections</span>
-        <span class="sidebar-stat-value">{st.total_detections}</span>
-    </div>
-    <div class="sidebar-stat">
-        <span class="sidebar-stat-label">Queue Length</span>
-        <span class="sidebar-stat-value">{len(st.global_gesture_queue)}</span>
-    </div>
-    <div class="sidebar-stat">
-        <span class="sidebar-stat-label">Uptime</span>
-        <span class="sidebar-stat-value">{uptime_str}</span>
-    </div>
-    <div class="sidebar-stat">
-        <span class="sidebar-stat-label">Model Status</span>
-        <span class="sidebar-stat-value">{"READY" if model else "MISSING"}</span>
-    </div>
-    """, unsafe_allow_html=True)
-
-    st.markdown("<br>", unsafe_allow_html=True)
-    st.markdown("**Supported Gestures**")
-
-    for emoji, name in [("🖐️", "Open Palm"), ("✊", "Closed Fist"), ("👍", "Thumbs Up"), ("✌️", "Peace Sign"), ("👌", "OK Sign")]:
-        st.markdown(f"""
-        <div class="sidebar-stat">
-            <span class="sidebar-stat-label">{emoji} {name}</span>
-            <span class="sidebar-stat-value">ACTIVE</span>
-        </div>
-        """, unsafe_allow_html=True)
-
-    st.markdown("<br>", unsafe_allow_html=True)
-    st.markdown("""
-    <p style="font-size:0.72rem; color:#475569; font-family:'Space Mono',monospace; line-height:1.6;">
-    MediaPipe hand landmarker feeds 21-point skeleton coordinates into a trained sklearn classifier. 
-    Anti-jitter latching prevents duplicate events across 4 stable frames.
-    </p>
-    """, unsafe_allow_html=True)
-
-
-# =========================================================
-# HERO HEADER
-# =========================================================
-st.markdown("""
-<div class="hero-wrapper">
-    <div class="hero-eyebrow">Real-Time Computer Vision</div>
-    <h1 class="hero-title">Gestura <span>AI</span></h1>
-    <p class="hero-sub">
-        Hand gesture recognition powered by MediaPipe landmarks and a trained
-        scikit-learn classifier — running live in your browser.
-    </p>
-</div>
-""", unsafe_allow_html=True)
-
-
-# =========================================================
-# MAIN LAYOUT — Two-Column Grid
-# =========================================================
-left_col, right_col = st.columns([1.1, 1.3], gap="large")
-
-# ── LEFT — Camera Feed ─────────────────────────────
-with left_col:
-    st.markdown('<div class="panel">', unsafe_allow_html=True)
-    st.markdown("""
-    <div style="display:flex;align-items:center;justify-content:space-between;margin-bottom:1rem;">
-        <span class="panel-title">Camera Input</span>
-        <span class="live-badge"><span class="live-dot"></span>Live</span>
-    </div>
-    """, unsafe_allow_html=True)
-
-    if model is not None:
-        webrtc_streamer(
-            key="gestura-engine",
-            video_processor_factory=GestureProcessor,
-            media_stream_constraints={"video": True, "audio": False},
-            async_processing=True,
-        )
-    else:
-        st.error(
-            "⚠️ Model assets not found. "
-            "Ensure `gesture_classifier.pkl` and `hand_landmarker.task` exist "
-            "inside `model/keypoint_classifier/`."
-        )
-
-    st.markdown("</div>", unsafe_allow_html=True)
-    st.markdown("<br>", unsafe_allow_html=True)
-
-    if st.button("↻  Refresh Dashboard", use_container_width=True):
-        st.rerun()
-
-
-# ── RIGHT — Analytics Panel ────────────────────────
-with right_col:
-    st.markdown('<div class="panel">', unsafe_allow_html=True)
-    st.markdown('<div class="panel-title">Detection Output</div>', unsafe_allow_html=True)
-
-    # — Current Gesture Widget —
-    # Read pre-split values written directly by the processor — no string parsing needed
-    cur_emoji = getattr(st, "latest_gesture_emoji", "🫥")
-    cur_label = getattr(st, "latest_gesture_label", "No Hand Detected")
-
-    st.markdown(f"""
-    <div class="gesture-display">
-        <div class="gesture-emoji">{cur_emoji}</div>
-        <div class="gesture-label">{cur_label}</div>
-        <div class="gesture-sub">current active detection</div>
-    </div>
-    """, unsafe_allow_html=True)
-
-    st.markdown("<hr>", unsafe_allow_html=True)
-
-    # — Tabs —
-    tab_options = ["📋 Gesture Log", "📊 Analytics", "📖 Key Map"]
-    selected_tab = st.radio("Panel View", tab_options, horizontal=True, label_visibility="collapsed")
-
-    st.markdown("<br>", unsafe_allow_html=True)
-
-    # ── TAB 1: Gesture Log ──────────────────────────
-    if selected_tab == "📋 Gesture Log":
-        st.markdown('<div class="panel-title">Sequential Detection Log</div>', unsafe_allow_html=True)
-
-        if len(st.global_gesture_queue) == 0:
-            st.markdown("""
-            <div style="text-align:center;padding:2rem;color:#475569;font-family:'Space Mono',monospace;font-size:0.8rem;">
-                — No gestures captured yet —<br>
-                <span style="font-size:0.7rem;opacity:0.6;">Show your hand to the camera to begin</span>
-            </div>
-            """, unsafe_allow_html=True)
-        else:
-            queue_reversed = list(reversed(st.global_gesture_queue))
-            log_html = ""
-            for i, gesture in enumerate(queue_reversed):
-                idx_label = len(st.global_gesture_queue) - i
-                new_tag   = '<span class="log-new-tag">new</span>' if i == 0 else ""
-                log_html += f"""
-                <div class="log-entry">
-                    <span class="log-index">#{idx_label:03d}</span>
-                    <span class="log-gesture">{gesture}</span>
-                    {new_tag}
-                </div>
-                """
-            st.markdown(log_html, unsafe_allow_html=True)
-
-        st.markdown("<br>", unsafe_allow_html=True)
-        if st.button("🗑  Clear Log & Reset Session", use_container_width=True):
-            st.global_gesture_queue      = []
-            st.latest_detected_gesture   = "No Hand Detected"
-            st.latest_gesture_emoji      = "🫥"
-            st.latest_gesture_label      = "No Hand Detected"
-            st.total_detections          = 0
-            st.session_start             = time.time()
-            st.rerun()
-
-    # ── TAB 2: Analytics ────────────────────────────
-    elif selected_tab == "📊 Analytics":
-        st.markdown('<div class="panel-title">System Performance</div>', unsafe_allow_html=True)
-        st.markdown("""
-        <div class="metric-grid">
-            <div class="metric-card">
-                <div class="metric-label">Target Latency</div>
-                <div class="metric-value">12<span style="font-size:1rem;color:#475569;">ms</span></div>
-                <div class="metric-tag">⚡ Sub-16ms rendering</div>
-            </div>
-            <div class="metric-card">
-                <div class="metric-label">Model Accuracy</div>
-                <div class="metric-value">98.4<span style="font-size:1rem;color:#475569;">%</span></div>
-                <div class="metric-tag">🧠 5-class classifier</div>
-            </div>
-            <div class="metric-card">
-                <div class="metric-label">Landmark Points</div>
-                <div class="metric-value">21</div>
-                <div class="metric-tag">🖐️ Per hand skeleton</div>
-            </div>
-            <div class="metric-card">
-                <div class="metric-label">Stability Gate</div>
-                <div class="metric-value">4<span style="font-size:1rem;color:#475569;">fr</span></div>
-                <div class="metric-tag">🔒 Anti-jitter latch</div>
-            </div>
-        </div>
-        """, unsafe_allow_html=True)
-
-        st.markdown("<br>", unsafe_allow_html=True)
-        st.markdown('<div class="panel-title">Pipeline Architecture</div>', unsafe_allow_html=True)
-        st.markdown("""
-        <div style="font-family:'Space Mono',monospace;font-size:0.72rem;color:#475569;line-height:2;padding:1rem;background:#111c30;border-radius:10px;border:1px solid rgba(255,255,255,0.06);">
-            Camera Frame<br>
-            &nbsp;&nbsp;↓ flip + resize (640×480)<br>
-            MediaPipe HandLandmarker<br>
-            &nbsp;&nbsp;↓ 21 × (x,y) keypoints<br>
-            Normalize relative to wrist<br>
-            &nbsp;&nbsp;↓ 42-dim feature vector<br>
-            Sklearn Classifier<br>
-            &nbsp;&nbsp;↓ 5-class prediction<br>
-            Anti-jitter latch (4 frames)<br>
-            &nbsp;&nbsp;↓ stable gesture event<br>
-            Dashboard Log
-        </div>
-        """, unsafe_allow_html=True)
-
-    # ── TAB 3: Key Map ───────────────────────────────
-    elif selected_tab == "📖 Key Map":
-        st.markdown('<div class="panel-title">Supported Gesture Reference</div>', unsafe_allow_html=True)
-
-        keymap = [
-            ("🖐️", "Open Palm",   "All five fingers extended and spread"),
-            ("✊", "Closed Fist", "All fingers curled into a closed fist"),
-            ("👍", "Thumbs Up",   "Thumb extended upward, fingers closed"),
-            ("✌️", "Peace Sign",  "Index and middle fingers raised in V"),
-            ("👌", "OK Sign",     "Thumb and index finger form a circle"),
-        ]
-
-        for emoji, name, desc in keymap:
-            st.markdown(f"""
-            <div class="keymap-row">
-                <div class="keymap-sign">{emoji}</div>
-                <div>
-                    <div class="keymap-name">{name}</div>
-                    <div class="keymap-desc">{desc}</div>
-                </div>
-            </div>
-            """, unsafe_allow_html=True)
-
-        st.markdown("<br>", unsafe_allow_html=True)
-        st.markdown("""
-        <div style="font-family:'Space Mono',monospace;font-size:0.7rem;color:#475569;line-height:1.8;padding:1rem;background:#111c30;border-radius:10px;border:1px solid rgba(255,255,255,0.06);">
-            <strong style="color:#94a3b8;">How detection works</strong><br>
-            Each hand is reduced to 21 landmark coordinates.<br>
-            Coordinates are normalized relative to the wrist<br>
-            and scaled to unit range before classification.<br>
-            A new event fires only after 4 consecutive<br>
-            stable frames of the same prediction.
-        </div>
-        """, unsafe_allow_html=True)
-
-    st.markdown("</div>", unsafe_allow_html=True)
